+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.20.1] - 2023-06-9
+
+### Changed
+
+- Updated user card styles in user management screen
+
 ## [1.20.0] - 2023-06-27
 
 ### Added
