+++ src/pages/users.tsx
@@ -1,12 +1,11 @@
 import type { NextPage, GetServerSideProps } from "next";
-import { useRouter } from "next/router";
 import Head from "next/head";
 import { getServerSession } from "next-auth";
 import { useState, ChangeEvent } from "react";
 import { authOptions } from "./api/auth/[...nextauth]";
 import { listUsers } from "./api/user/list";
 import { User } from "../types";
-import { Badge } from "../components";
+import { UserCard } from "../components/UserCard";
 
 type Props = {
   users: User[];
@@ -14,8 +13,6 @@ type Props = {
 
 const UsersList: NextPage<Props> = ({ users }: Props) => {
   const [searchText, setSearchText] = useState<string | null>(null);
-  const [isSwiping, setSwiping] = useState(false);
-  const router = useRouter();
 
   if (!users) {
     return null; // Redirect? or error
@@ -23,10 +20,6 @@ const UsersList: NextPage<Props> = ({ users }: Props) => {
 
   const handleSearch = (event: ChangeEvent<HTMLInputElement>) => {
     setSearchText(event.currentTarget.value);
-  };
-
-  const onPress = (id: string) => {
-    router.push(`/profile/${id}`);
   };
 
   const filteredUsers = searchText
@@ -58,41 +51,8 @@ const UsersList: NextPage<Props> = ({ users }: Props) => {
       </div>
 
       <div className="grid w-full grid-cols-1 gap-2 md:gap-2 px-2 sm:px-0">
-        {filteredUsers.map(({ id, name, email, role }) => (
-          <div
-            key={id}
-            id={id}
-            className="md:mx-autotext-neutral-500 box-border flex w-full cursor-pointer gap-2 rounded bg-white shadow-md hover:text-neutral-700 hover:shadow-lg md:gap-2"
-            role="presentation"
-            onMouseDown={() => setSwiping(false)}
-            onMouseMove={() => setSwiping(true)}
-            onMouseUp={(e) => {
-              if (!isSwiping && e.button === 0) {
-                onPress(id);
-              }
-              setSwiping(false);
-            }}
-            onTouchStart={() => setSwiping(false)}
-            onTouchMove={() => setSwiping(true)}
-            onTouchEnd={(e) => {
-              if (e.cancelable) e.preventDefault();
-              if (!isSwiping) {
-                onPress(id);
-              }
-              setSwiping(false);
-            }}
-          >
-            <div className="flex-auto align-middle p-1 px-2 font-bold uppercase tracking-wide">
-              {name}
-            </div>
-            {["ADMIN", "LEADER"].includes(role) && (
-              <div className="pt-[3px]">
-                <Badge text={role} />
-              </div>
-            )}
-
-            <div className="truncate p-1">{email}</div>
-          </div>
+        {filteredUsers.map((user) => (
+          <UserCard key={user.id} user={user} />
         ))}
       </div>
     </>
